moved cli message composer functions out into new module cli_composer.go
